Move `RejectionType` under the `type` package
